+++ supabase/functions/analyze_countries_policies/types.ts
@@ -1,6 +1,18 @@
 
-export { 
-  Country, 
-  ProcessingError, 
-  CountryPolicyResult 
-} from '../_shared/types.ts';
+export interface Country {
+  id: string;
+  name: string;
+  code: string;
+}
+
+export interface ProcessingError {
+  country: string;
+  error: string;
+}
+
+export interface CountryPolicyResult {
+  country: string;
+  processingTimeMs: number;
+  policiesFound: number;
+  policiesSaved?: string[];
+}
